switch to pathname "Quick to Read/Title Case"
The previous format is like "foobar/LikeThis.txt," which is fine, but
when one is referred from a text file, the readability is not as good as
one with spaces. For a quick comparison,

  [...] a medical documentary like TV Shows/24 Hours in A&E might [...]
  [...] a medical documentary like TV/shows/24HoursInA&E.txt might [...]

Basically, everything is Title Cased, including the directory names,
which were not. All more consistent and more human-readable.

This format is not good for machine, but there is no plan for coding a
reader, and probably never would have.